Split ColorData from StyleData into it's own file
