+++ include/SSVOpenHexagon/Data/StyleData.hpp
@@ -5,6 +5,7 @@
 #pragma once
 
 #include "SSVOpenHexagon/Global/Common.hpp"
+#include "SSVOpenHexagon/Data/ColorData.hpp"
 #include "SSVOpenHexagon/Data/CapColor.hpp"
 
 namespace hg
@@ -15,29 +16,6 @@ struct LevelStatus;
 class StyleData
 {
 private:
-    struct ColorData
-    {
-        bool main, dynamic, dynamicOffset;
-        float dynamicDarkness, hueShift, offset;
-        sf::Color color, pulse;
-
-        ColorData() = default;
-        ColorData(const ssvuj::Obj& mRoot)
-            : main{ssvuj::getExtr<bool>(mRoot, "main", false)},
-              dynamic{ssvuj::getExtr<bool>(mRoot, "dynamic", false)},
-              dynamicOffset{
-                  ssvuj::getExtr<bool>(mRoot, "dynamic_offset", false)},
-              dynamicDarkness{
-                  ssvuj::getExtr<float>(mRoot, "dynamic_darkness", 1.f)},
-              hueShift{ssvuj::getExtr<float>(mRoot, "hue_shift", 0.f)},
-              offset{ssvuj::getExtr<float>(mRoot, "offset", 0.f)},
-              color{
-                  ssvuj::getExtr<sf::Color>(mRoot, "value", sf::Color::White)},
-              pulse{ssvuj::getExtr<sf::Color>(mRoot, "pulse", sf::Color::White)}
-        {
-        }
-    };
-
     float currentHue, currentSwapTime{0}, pulseFactor{0};
     Path rootPath;
     sf::Color currentMainColor, current3DOverrideColor;
